feat(Turnips): add Turnips chart part 5
* Start splitting the big TurnipsChartView into more smaller and readable files
